Updated bunch of stuff

+++ src/environment_typing.py
@@ -5,7 +5,7 @@ import numpy as np
 
 class envt:
     uint = np.uint
-    int = np.int32
-    float = np.float32
+    int = np.int64
+    float = np.float64
 
     Any = typing.Any
